moving data collector to MetalContext
